Rename `circuit_drawer` module to `drawer` (#1949)
* rename circuit_drawer module to drawer

* black

* Update doc/releases/changelog-dev.md
