--- doc/code/qml_circuit_drawer.rst
@@ -1,16 +1,0 @@
-qml.circuit_drawer
-==================
-
-.. currentmodule:: pennylane.circuit_drawer
-
-.. warning::
-
-    Unless you are a PennyLane or plugin developer, you likely do not need
-    to use these classes directly.
-
-    The :class:`~pennylane.circuit_drawer.MPLDrawer` can be used manually but does not yet integrate with the rest of PennyLane.
-
-.. automodapi:: pennylane.circuit_drawer
-    :no-heading:
-    :no-inheritance-diagram:
-    :no-inherited-members:
