+++ doc/code/qml_drawer.rst
@@ -1,0 +1,14 @@
+qml.drawer
+==========
+
+.. currentmodule:: pennylane.drawer
+
+.. warning::
+
+    Unless you are a PennyLane or plugin developer, you likely do not need
+    to use these classes directly.
+
+.. automodapi:: pennylane.drawer
+    :no-heading:
+    :no-inheritance-diagram:
+    :no-inherited-members:
