+++ doc/index.rst
@@ -226,7 +226,7 @@ PennyLane is **free** and **open source**, released under the Apache License, Ve
    code/qml_qnn
    code/qml_tape
    code/qml_transforms
-   code/qml_circuit_drawer
+   code/qml_drawer
    code/qml_utils
    code/qml_wires
    code/qml_beta
