+++ doc/releases/changelog-dev.md
@@ -354,6 +354,9 @@
 
 <h3>Breaking changes</h3>
 
+* The `circuit_drawer` module has been renamed `drawer`.
+  [(#1949)](https://github.com/PennyLaneAI/pennylane/pull/1949)
+
 * The `par_domain` attribute in the operator class has been removed.
   [(#1907)](https://github.com/PennyLaneAI/pennylane/pull/1907)
