+++ pennylane/drawer/representation_resolver.py
