+++ tests/drawer/test_representation_resolver.py
@@ -12,14 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """
-Unit tests for the :mod:`pennylane.circuit_drawer.representation_resolver` module.
+Unit tests for the :mod:`pennylane.drawer.representation_resolver` module.
 """
 from unittest.mock import Mock
 import pytest
 import numpy as np
 
 import pennylane as qml
-from pennylane.circuit_drawer.representation_resolver import RepresentationResolver
+from pennylane.drawer.representation_resolver import RepresentationResolver
 from pennylane.measure import state
 
 
@@ -32,7 +32,7 @@ def unicode_representation_resolver():
 @pytest.fixture
 def ascii_representation_resolver():
     """An instance of a RepresentationResolver with unicode charset."""
-    return RepresentationResolver(charset=qml.circuit_drawer.charsets.AsciiCharSet)
+    return RepresentationResolver(charset=qml.drawer.charsets.AsciiCharSet)
 
 
 class TestRepresentationResolver:
